ci: Don't remove dir

+++ vendor-one.sh
@@ -79,4 +79,6 @@ git add .
   git -C "$upstream_dir" log --first-parent --format="%s" ${base}..${commit} | tee /dev/stderr | sed -r 's%(#[0-9]+)%duckdb/duckdb\1%g'
 ) | git commit --file /dev/stdin
 
-rm -rf "$upstream_dir"
+if [ "$upstream_basedir" != ".git/duckdb" ]; then
+  rm -rf "$upstream_dir"
+fi
